+++ src/pflow/core/prompt_cache_analysis/__init__.py
@@ -137,10 +137,11 @@ from __future__ import annotations
 
 from typing import Final
 
-from .analyze import analyze, list_traces_for_workflow
+from .analyze import analyze
 from .render_json import render_json
 from .render_text import render_text
 from .summarize import summarize, summarize_from_analysis
+from .trace_loading import list_traces_for_workflow
 from .types import CacheAnalysis, TraceListEntry
 
 JSON_FORMAT_VERSION: Final[str] = "5.0"
